completed debugging of the new subset function: The error was that i thought of the data region as limited by the start point and the last difference, when in fact the maximum order space was defined as the maximum difference not including the last difference in range, since the initial differences were initialised with the last diff being NAN.

+++ src/qalgorithms_input_output.cpp
@@ -728,7 +728,7 @@ namespace qAlgorithms
         for (size_t i = 0; i < peaktable->size(); i++)
         {
             const FeaturePeak peak = peaktable->at(i);
-            int binID = peak.idxBin;
+            size_t binID = peak.idxBin;
             assert(binID < originalBins->size());
             const std::vector<unsigned int> *scanNums = &(originalBins->at(binID).scanNumbers);
             assert(peak.idxBinStart < peak.idxBinEnd);
